test(e2e): simplify E2E tests into 2 files
- Consolidate home/records/settings tests into app.test.ts
- Simplify record-flows.test.ts with shared verification helper
- Reduce from 18 tests in 4 files to 9 tests in 2 files
