+++ src/types/index.ts
@@ -12,7 +12,6 @@ export interface BaseRecord {
   userId: string;
   date: string; // 2026-04-10
   time: string; // 08:30
-  note?: string;
   createdAt: Date;
   deletedAt?: Date;
 }
@@ -22,22 +21,26 @@ export interface SymptomRecord extends BaseRecord {
   symptoms: string[]; // 症状列表
   severity?: 1 | 2 | 3; // 整体严重程度：轻度、中度、重度
   overallFeeling: 1 | 2 | 3 | 4 | 5; // 整体感受 1很差 - 5很好
+  note?: string;
 }
 
 // 饮食记录
 export interface MealRecord extends BaseRecord {
   foods: string[];
   amount: 1 | 2 | 3; // 1-少量 2-适中 3-大量
+  note?: string;
 }
 
 // 排便记录
 export interface StoolRecord extends BaseRecord {
   type: 1 | 2 | 3 | 4 | 5 | 6 | 7; // Bristol 分类
   amount: 1 | 2 | 3; // 1-少量 2-适中 3-大量
+  note?: string;
 }
 
 // 用药记录
 export interface MedicationRecord extends BaseRecord {
   name: string;
   dosage: string;
+  note?: string;
 }
